fix: harden pipeline plan validation tests
